Refactor nacos task execute engine
