+++ config/src/main/java/com/alibaba/nacos/config/server/service/notify/NotifySingleService.java
@@ -18,7 +18,7 @@ package com.alibaba.nacos.config.server.service.notify;
 
 import com.alibaba.nacos.common.executor.ExecutorFactory;
 import com.alibaba.nacos.common.executor.NameThreadFactory;
-import com.alibaba.nacos.common.task.AbstractDelayTask;
+import com.alibaba.nacos.common.task.NacosTask;
 import com.alibaba.nacos.config.server.utils.LogUtil;
 import com.alibaba.nacos.core.cluster.Member;
 import com.alibaba.nacos.core.cluster.ServerMemberManager;
@@ -50,7 +50,7 @@ public class NotifySingleService {
         }
         
         @Override
-        public boolean process(AbstractDelayTask task) {
+        public boolean process(NacosTask task) {
             NotifySingleTask notifyTask = (NotifySingleTask) task;
             return notifyToDump(notifyTask.getDataId(), notifyTask.getGroup(), notifyTask.getTenant(),
                     notifyTask.getLastModified(), notifyTask.target);
