+++ config/src/main/java/com/alibaba/nacos/config/server/service/notify/NotifyTaskProcessor.java
@@ -17,8 +17,8 @@
 package com.alibaba.nacos.config.server.service.notify;
 
 import com.alibaba.nacos.common.model.RestResult;
+import com.alibaba.nacos.common.task.NacosTask;
 import com.alibaba.nacos.config.server.constant.Constants;
-import com.alibaba.nacos.common.task.AbstractDelayTask;
 import com.alibaba.nacos.common.task.NacosTaskProcessor;
 import com.alibaba.nacos.config.server.monitor.MetricsMonitor;
 import com.alibaba.nacos.config.server.service.trace.ConfigTraceService;
@@ -46,7 +46,7 @@ public class NotifyTaskProcessor implements NacosTaskProcessor {
     }
     
     @Override
-    public boolean process(AbstractDelayTask task) {
+    public boolean process(NacosTask task) {
         NotifyTask notifyTask = (NotifyTask) task;
         String dataId = notifyTask.getDataId();
         String group = notifyTask.getGroup();
